Fix usage of cli_and_save/restore_flags
We should NOT be calling sti() after we restore flags - this completely
defeats the purpose of saving the flags. This caused interrupts to be
enabled whenever enable_irq/rtc_init was being called, which we don't want.

Also cleaned up bootup script, and placed it in between the interrupt
blocking time so that the user cannot enter keypresses at the boot screen
and mess with the keyboard buffer before the system has fully booted
past the bootscreen.

Signed-off-by: Rohan Mathur <rohan@rmathur.com>