Clean up syscall flow + vidmap warnings
Originally, we had to sti at the beginning of terminal read, and I wasn't
really sure why. Found out it was because we are setting up the syscall
IDT entry as an interrupt, which automatically clears the interrupt flag
when it enters that vector. STIing in a more general case in the syscall
asm wrapper is a much better solution since it enables interrupts for
all syscalls and not just after terminal_read() is called.

Signed-off-by: Rohan Mathur <rohan@rmathur.com>